chore(nlu): move back nlu definitions inside nlu module
