+++ modules/basic-skills/src/actions/slot_update_contexts.js
@@ -8,7 +8,7 @@ const axios = require('axios')
 const updateContexts = async intentName => {
   const botId = event.botId
   const axiosConfig = await bp.http.getAxiosConfigForBot(botId, { localUrl: true })
-  const { data } = await axios.get(`/nlu/intents/${intentName}`, axiosConfig)
+  const { data } = await axios.get(`/mod/nlu/intents/${intentName}`, axiosConfig)
 
   const nluContexts = data.contexts.map(context => {
     return {
