+++ modules/basic-skills/src/views/full/slot.jsx
@@ -69,7 +69,7 @@ export class Slot extends React.Component {
   }
 
   fetchIntents = () => {
-    return this.props.bp.axios.get('/nlu/intents').then(({ data }) => {
+    return this.props.bp.axios.get('/mod/nlu/intents').then(({ data }) => {
       const intents = data.filter(x => !x.name.startsWith('__qna'))
       this.setState({ intents })
     })
