+++ modules/nlu-testing/src/views/full/api.ts
@@ -21,7 +21,7 @@ export const makeApi = (bp: { axios: AxiosInstance }): TestingAPI => {
     },
 
     fetchIntents: async () => {
-      const { data } = await bp.axios.get('/nlu/intents')
+      const { data } = await bp.axios.get('/mod/nlu/intents')
       return data
     },
