+++ modules/nlu/src/backend/definitions/utils.ts
@@ -1,0 +1,6 @@
+export const sanitizeFileName = (name: string): string => {
+  return name
+    .toLowerCase()
+    .replace(/\.json$/i, '')
+    .replace(/[\t\s]/gi, '-')
+}
