--- modules/nlu/src/backend/entities/validation.ts
@@ -1,5 +1,0 @@
-export const FuzzyTolerance = {
-  Loose: 0.65,
-  Medium: 0.8,
-  Strict: 1
-}
