+++ modules/nlu/src/views/full/entities/EntityNameModal.tsx
@@ -4,7 +4,7 @@ import { lang } from 'botpress/shared'
 import _ from 'lodash'
 import React, { FC, useEffect, useState } from 'react'
 
-import { NLUApi } from '../../../api'
+import { NLUApi } from '../../../api-client'
 
 const AVAILABLE_TYPES = [
   {
