+++ src/bp/core/server.ts
@@ -37,7 +37,6 @@ import { AdminRouter, AuthRouter, BotsRouter, MediaRouter, ModulesRouter } from 
 import { ContentRouter } from './routers/bots/content'
 import { ConverseRouter } from './routers/bots/converse'
 import { HintsRouter } from './routers/bots/hints'
-import { NLURouter } from './routers/bots/nlu'
 import { isDisabled } from './routers/conditionalMiddleware'
 import { InvalidExternalToken, PaymentRequiredError } from './routers/errors'
 import { SdkApiRouter } from './routers/sdk/router'
@@ -60,7 +59,6 @@ import { HintsService } from './services/hints'
 import { JobService } from './services/job-service'
 import { MediaServiceProvider } from './services/media'
 import { MonitoringService } from './services/monitoring'
-import { NLUService } from './services/nlu/nlu-service'
 import { NotificationsService } from './services/notification/service'
 import { WorkspaceService } from './services/workspace-service'
 import { TYPES } from './types'
@@ -91,7 +89,6 @@ export default class HTTPServer {
   private readonly adminRouter: AdminRouter
   private readonly botsRouter: BotsRouter
   private contentRouter!: ContentRouter
-  private nluRouter!: NLURouter
   private readonly modulesRouter: ModulesRouter
   private readonly shortLinksRouter: ShortLinksRouter
   private converseRouter!: ConverseRouter
@@ -140,7 +137,6 @@ export default class HTTPServer {
     @inject(TYPES.AlertingService) private alertingService: AlertingService,
     @inject(TYPES.JobService) private jobService: JobService,
     @inject(TYPES.LogsRepository) private logsRepo: LogsRepository,
-    @inject(TYPES.NLUService) private nluService: NLUService,
     @inject(TYPES.TelemetryRepository) private telemetryRepo: TelemetryRepository
   ) {
     this.app = express()
@@ -252,12 +248,10 @@ export default class HTTPServer {
       this.workspaceService,
       this.ghostService
     )
-    this.nluRouter = new NLURouter(this.logger, this.authService, this.workspaceService, this.nluService)
     this.converseRouter = new ConverseRouter(this.logger, this.converseService, this.authService, this)
     this.hintsRouter = new HintsRouter(this.logger, this.hintsService, this.authService, this.workspaceService)
     this.botsRouter.router.use('/content', this.contentRouter.router)
     this.botsRouter.router.use('/converse', this.converseRouter.router)
-    this.botsRouter.router.use('/nlu', this.nluRouter.router)
 
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     AppLifecycle.waitFor(AppLifecycleEvents.BOTPRESS_READY).then(() => {
