+++ src/bp/ui-studio/src/web/actions/index.ts
@@ -429,7 +429,7 @@ export const refreshActions = () => dispatch => {
 export const intentsReceived = createAction('INTENTS/RECEIVED')
 export const refreshIntents = () => dispatch => {
   // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  axios.get(`${window.BOT_API_PATH}/nlu/intents`).then(({ data }) => {
+  axios.get(`${window.BOT_API_PATH}/mod/nlu/intents`).then(({ data }) => {
     dispatch(intentsReceived(data))
   })
 }
